Created file for training modd2 with detectron2
I created a file to train the modd2 dataset using detectron2.
There are issues with the file, and they will have to be addressed
to make it more robust. It's also in the process of becoming more
modular, but that requires more time that I don't have right now.
I have started a requirements.txt file as well that I am sure is
incomplete. I have also updated data_loading.py to match the
get_data() function inside of train.py