Created custom evaluator and flexible evaluation metric with COCOAPI source code
I adapted the COCOEvaluator and COCOEval classes from both detectron2
and pycocotools in order to evaluate our model using statistics similar
to the COCO evaluation, but with flexibility in the evaluation parameters.
Things like iou thresholds can be changed and modified to personal needs,
such as our need to evaluate our model on a similar level to the MODD2
data evaluation method.

Added a CounterEvaluator which is a copy of the example code of the
custom evaluators in detectron2.

The baseline_800iter...txt file is the output of the evaluation on
our baseline model with no data augmentation for the training set.